Update folder structure and code format
